refactor: store domain package refactoring
